Introduce ORM via SQLModel
This brings in the SQLModel ORM system (which builds on
SQLAlchemy). The attempt is not to change the BraidDB class
programming model used by the various workflow tests, but we may want
to relax this over time to provide a somewhat more flexible query
model which takes advantage of the ORM model now available underlying
the BraidDB.

This also probably breaks the MPI work. Will have to update that
accordingly, but this would also probably require work once new query
models are introduced, so it is worth getting queries back in place
and then address the MPI when they are there and when we have a need
for the MPI interface.