Add invalidation action
This adds an invalidation action that can be associated with any
record. When the record is marked invalid, the invalidation action
will be executed.

At present, the only invalidation action type supported is executing a
subprocess. In longer term, generating an Event to the Globus Automate
infrastructure may be supported so that invalidation of a record could
cause other Automate actions to occur.