Add capability to register and invoke braid functions via funcx.
At present, only one method is exposed which can be used to add a
record to the braid database. Undoubtedly, additional methods for
doing bulk addition of records, lookup, update and invalidation
operations will be needed.

Read the funcx section of README.adoc to see how to configure and
otherwise setup for working with funcx.